Moved OnSeries mixin as OnSeriesComposition.
